Исправил недочёты

+++ MetricsAgent/DAL/HddMetricsRepository.cs
@@ -18,7 +18,6 @@ namespace MetricsAgent.DAL
             using var connection = new SQLiteConnection(ConnectionString);
             connection.Open();
             using var cmd = new SQLiteCommand(connection);
-
             string comText = $"SELECT * FROM metrics WHERE (time > {fromTime.ToUnixTimeSeconds()}) AND (time < {toTime.ToUnixTimeSeconds()})";
             cmd.CommandText = comText;
             cmd.ExecuteNonQuery();
@@ -32,8 +31,8 @@ namespace MetricsAgent.DAL
                     returnList.Add(new HddMetric
                     {
                         Id = reader.GetInt32(0),
-                        Value = reader.GetInt32(1),
-                        Time = DateTimeOffset.FromUnixTimeSeconds(reader.GetInt32(2))
+                        Value = reader.GetInt64(1),
+                        Time = DateTimeOffset.FromUnixTimeSeconds(reader.GetInt64(2))
                     });
                 }
             }
